chore(baseline): stabilize build and tests across packages
Fixes accumulated during baseline triage so all listed packages build
cleanly and the util module test suite is fully green, unblocking
sprint execution.

util:
- BitSetIterator.DocIDRunEnd advances iterator to last set bit in run
  (matches Lucene semantics; downstream NextDoc() now resumes correctly)
- NaN normalization in DoubleToSortableLong/FloatToSortableInt so all
  NaN bit patterns map to canonical positive NaN (matches Java
  Double.doubleToLongBits)
- BigIntToSortableBytes: positive values whose first magnitude byte has
  the MSB set now allocate an extra sign byte, matching Java
  BigInteger.toByteArray
- SortableBytesToBigInt: detect original sign from ENCODED MSB before
  flipping (was reading sign from already-flipped buffer)
- TestDoublesRoundTrip / TestFloatsRoundTrip use IsNaN-aware comparison
- ApplyToAll returns the first error (not a wrapped aggregate),
  matching Lucene IOUtils.applyToAll semantics

analysis:
- CachingTokenFilter no longer double-emits: when the input stream is
  exhausted on the first pass, return false rather than falling through
  to the cache (callers must Reset() to replay)
- KeywordRepeatFilter registers KeywordAttribute on the shared
  AttributeSource when missing, so upstream tokenizers without that
  attribute (e.g. WhitespaceTokenizer) no longer cause nil dereferences
- analyzers_extended_test: WhitespaceAnalyzer emits one token for a run
  of punctuation (it splits on whitespace only), expectation corrected
- ascii_folding_filter_test: Æ folds to "AE" not "A", expectation
  corrected ("Æther" -> "AEther")

queryparser:
- Boolean operator detection runs AFTER field detection: words like
  "title" no longer get pre-empted by the A/O/N/T-prefix operator check
  and "field:term" parses correctly
- createTermQuery applies the configured analyzer so "Hello" indexes as
  "hello" under StandardAnalyzer; multi-token expansions yield a SHOULD
  BooleanQuery
- parseAndExpression / parseExpression flatten consecutive operands
  into a single BooleanQuery (e.g. "a AND b AND c" -> 3 MUST clauses)

Plus prior package-level cleanup from the in-progress baseline pass:
test files updated to match current APIs across codecs/document/facets/
index/join/memory/queryparser/search/spatial/store/suggest.

+++ search/top_field_collector_test.go
@@ -451,7 +451,7 @@ func TestTopFieldCollector_RelationVsTopDocsCount(t *testing.T) {
 		if expectedLen > 10 {
 			expectedLen = 10
 		}
-		if int64(len(topDocs.ScoreDocs)) != expectedLen {
+		if len(topDocs.ScoreDocs) != expectedLen {
 			t.Errorf("Expected %d score docs with numHits=%d, got %d", expectedLen, tc.numHits, len(topDocs.ScoreDocs))
 		}
 	}
